requirements.txt
# Core dependencies
anthropic>=0.18.0
click>=8.1.0
rich>=13.0.0
python-dotenv>=1.0.0

# Date parsing
python-dateutil>=2.8.2

# Development and testing
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
